Extract scene material plan composer
