+++ src/PlateauResoniteLink/Targets/Resonite/ResonitePackageSemantics.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal static class ResonitePackageSemantics
+{
+    public static bool IsDemPackage(string packageName)
+    {
+        return string.Equals(packageName, "dem", StringComparison.OrdinalIgnoreCase);
+    }
+}
